null check on inhoud field

+++ src/OneGround.ZGW.Documenten.Web/Handlers/v1/CreateEnkelvoudigInformatieObjectCommandHandler.cs
@@ -355,6 +355,11 @@ class CreateEnkelvoudigInformatieObjectCommandHandler
 
     private async Task RollbackDocumentJustAddedToDmsAsync(string urnDocument)
     {
+        if (string.IsNullOrEmpty(urnDocument))
+        {
+            return;
+        }
+
         try
         {
             await DocumentService.DeleteDocumentAsync(new DocumentUrn(urnDocument));
